Fix the MSVC build
This fork built only under MinGW/GCC; the checked-in MSVC project had
drifted. Portability fixes, no behaviour change under GCC:

* byteptr.h: the string/mem macros used GCC statement expressions
  (`({ ... })`). Converted to `do { ... } while (0)`, which works in both
  compilers and -- unlike upstream's bare `{ ... }` -- is still safe in an
  unbraced if/else. The value-returning READ* macros already had a
  non-GNUC fallback.
* doomtype.h / blua/luaconf.h: both #define'd INT32 (and friends) as
  macros. Windows' <basetsd.h> declares `typedef signed int INT32,
  *PINT32;`, so any TU reaching it afterwards failed to compile. Windows
  supplies the typedefs, so drop the macros; luaconf.h now uses int32_t
  for LUA_INTEGER/LUA_NUMBER/LUAI_UACNUMBER.
* deh_soc.c, d_netcmd.c: replaced C99 VLAs with fixed-size arrays. The
  bounds were already constant in practice (MAXCOLORNAME+1, MAX_ARGS);
  MAX_ARGS moved to command.h so callers can size arrays with it.
* r_things.c: `?: ` (GCC binary conditional) expanded to an if.
* hw_main.c, r_splats.c, lua_mathlib.c, m_menu.c, m_easing.c: unary minus
  on unsigned operands (C4146) rewritten as `0 - x` / an explicit signed
  cast, preserving modular semantics.
* r_opengl.c: M_PI isn't in MSVC's <math.h>; use M_PIl as elsewhere in the
  file. Added explicit casts for narrowing to GLfloat.
* p_mobj.c: `return <void expr>;` in a void function.
* m_easing.h: pass a named empty separator rather than an empty macro
  argument (C4003).
* w_wad.h: declare W_GetNumForNameMusic.
* v_video.c, r_picformats.c: unused local; pointer MSVC can't prove is
  initialised because it doesn't know I_Error is no-return.
* Srb2SDL-vc10.vcxproj: add m_easing.c (its symbols were undefined at
  link); drop i_cdmus.c (no longer exists) and strcasestr.c (textually
  included by string.c, doesn't compile standalone).
* SRB2_common.props: BrowseInformation off. The legacy .bsc database is
  unused, costs a ~4 MB .sbr per TU, and its writes intermittently fail
  under parallel builds.

+++ src/m_menu.c
@@ -8358,7 +8358,7 @@ static void M_DrawLoadGameData(void)
 	if (vid.width != BASEVIDWIDTH*vid.dupx)
 		hsep = (hsep*vid.width)/(BASEVIDWIDTH*vid.dupx);
 
-	for (i = 2; prev_i; i = -(i + ((UINT32)i >> 31))) // draws from outwards in; 2, -2, 1, -1, 0
+	for (i = 2; prev_i; i = -(INT32)(i + ((UINT32)i >> 31))) // draws from outwards in; 2, -2, 1, -1, 0
 	{
 		prev_i = i;
 		savetodraw = (saveSlotSelected + i + numsaves)%numsaves;
